+++ .storybook/preview.js
@@ -1,13 +1,10 @@
+// Internals
 import '../src/styles/storybook.css';
 
-// https://storybook.js.org/docs/react/writing-stories/parameters#global-parameters
+/**
+ * @type {import('@storybook/addons').Parameters}
+ */
 export const parameters = {
-  // https://storybook.js.org/docs/react/essentials/actions#automatically-matching-args
-  actions: { argTypesRegex: '^on.*' },
-  options: {
-    storySort: (a, b) =>
-      a[1].kind === b[1].kind
-        ? 0
-        : a[1].id.localeCompare(b[1].id, undefined, { numeric: true }),
-  },
+  layout: 'centered',
+  viewMode: 'docs',
 };