+++ jest.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  globals: {
+    'ts-jest': {
+      tsConfig: 'tsconfig.test.json',
+    },
+  },
+  moduleNameMapper: {
+    '^@/(.*)$': '<rootDir>/src/$1',
+  },
+  setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
+  transformIgnorePatterns: ['<rootDir>/node_modules/(?!@square/web-sdk)'],
+};