+++ src/components/ApplePay/styles.ts
@@ -1,0 +1,12 @@
+// Dependencies
+import { styled } from '@stitches/react';
+
+export const ApplePayContainer = styled('div', {
+  ApplePayButtonStyle: 'black',
+  ApplePayButtonType: 'plain',
+  cursor: 'pointer',
+  display: 'inline-block',
+  height: 48,
+  WebkitAppearance: '-apple-pay-button',
+  width: '100%',
+});