+++ src/components/GiftCardInput/styles.ts
@@ -1,19 +1,17 @@
-// Vendor Modules
-import { keyframes } from '@emotion/react';
-import styled, { CSSObject } from '@emotion/styled';
+// Dependencies
+import { keyframes, styled } from '@stitches/react';
 
-const pulse = keyframes`
-  0%, 100% {
-    opacity: 1;
-  }
+const pulse = keyframes({
+  '0%, 100%': {
+    opacity: 1,
+  },
+  '50%': {
+    opacity: 0.5,
+  },
+});
 
-  50% {
-    opacity: .5;
-  }
-`;
-
-export const LoadingCard = styled.div({
-  animation: `${pulse} 2s cubic-bezier(0.4, 0, 0.6, 1) infinite`,
+export const LoadingCard = styled('div', {
+  animation: `${pulse()} 2s cubic-bezier(0.4, 0, 0.6, 1) infinite`,
   background: '#F3F4F6',
   borderRadius: 6,
   height: 50,
@@ -21,21 +19,26 @@ export const LoadingCard = styled.div({
   position: 'relative',
 });
 
-export const PayButton = styled.button<{
-  overrideStyles?: CSSObject;
-}>((props) => ({
-  background: 'rgba(0, 0, 0, 1)',
-  borderRadius: 6,
-  border: 'none',
-  color: 'white',
+export const PayButton = styled('button', {
+  backgroundColor: '#006aff',
+  borderRadius: 5,
+  boxShadow: 1,
+  color: '#fff',
   cursor: 'pointer',
-  display: 'block',
+  borderStyle: 'none',
   fontSize: 16,
-  lineHeight: `18px`,
-  padding: 16,
+  fontWeight: 500,
+  lineHeight: '24px',
+  outline: 'none',
+  padding: 12,
+  userSelect: 'none',
   width: '100%',
-  '&:hover': {
-    background: 'rgba(0, 0, 0, 0.85)',
+  '&:active': {
+    backgroundColor: 'rgb(0, 85, 204)',
   },
-  ...props.overrideStyles,
-}));
+  '&:disabled': {
+    backgroundColor: 'rgba(0, 0, 0, 0.05)',
+    color: 'rgba(0, 0, 0, 0.3)',
+    cursor: 'not-allowed',
+  },
+});