+++ src/contexts/index.tsx
@@ -1,23 +1,17 @@
-// Vendor Modules
-import {
+// Dependencies
+import * as React from 'react';
+import type {
   ChargeVerifyBuyerDetails,
-  payments,
+  Payments,
+  PaymentRequestOptions,
   StoreVerifyBuyerDetails,
   TokenResult,
+  VerifyBuyerResponseDetails,
 } from '@square/web-sdk';
-import * as React from 'react';
-import type { Payments, PaymentRequestOptions } from '@square/web-sdk';
 
 // Internals
-import NoLocationIdOrAppId from '../components/NoLocationIdOrAppId/NoLocationIdOrAppId';
-import { INITIAL_STATE_METHODS } from '../constants';
 import { useDynamicCallback } from '../hooks';
-import { methodsReducer } from '../reducers';
-import type {
-  ActionMethodReducer,
-  FormContextInterface,
-  MethodsSupported,
-} from '../@types';
+import type { FormContextInterface } from '../types';
 
 /**
  * Export the hook here so we avoid circular dependency
@@ -39,51 +33,51 @@ export const useForm = (): FormContextInterface => {
  * [customization](customization.md) page for usage details.
  */
 export const FormContext = React.createContext<FormContextInterface>({
-  ach: 'loading',
-  applePay: 'loading',
-  card: 'loading',
-  cashApp: 'loading',
-  googlePay: 'loading',
-  giftCard: 'loading',
-  formId: '',
-  payments: (null as unknown) as Payments,
-  dispatchMethods: (null as unknown) as React.Dispatch<ActionMethodReducer>,
   cardTokenizeResponseReceived: (null as unknown) as (
-    props: TokenResult
+    token: TokenResult,
+    verifiedBuyer?: VerifyBuyerResponseDetails | null
   ) => void,
   createPaymentRequest: (null as unknown) as PaymentRequestOptions,
+  formId: '',
+  payments: (null as unknown) as Payments,
 });
 
-interface ProviderProps {
-  applicationId: string;
-  locationId: string;
+export interface ProviderProps {
   createPaymentRequest?: () => PaymentRequestOptions;
-  methodsSupported?: MethodsSupported;
-  cardTokenizeResponseReceived: (props: TokenResult) => void;
+  cardTokenizeResponseReceived: (
+    token: TokenResult,
+    verifiedBuyer?: VerifyBuyerResponseDetails | null
+  ) => void;
   createVerificationDetails?: () =>
     | ChargeVerifyBuyerDetails
     | StoreVerifyBuyerDetails;
+  payments: Payments | null;
 }
 
-const FormProvider: React.FC<ProviderProps> = ({ children, ...props }) => {
-  const [applicationId] = React.useState(() => props.applicationId);
-  const [locationId] = React.useState(() => props.locationId);
-  const [pay, setPay] = React.useState<Payments>();
+const FormProvider: React.FC<ProviderProps> = ({
+  children,
+  payments,
+  ...props
+}) => {
   const [createPaymentRequest] = React.useState<
     undefined | PaymentRequestOptions
   >(() => props.createPaymentRequest?.());
-  const [methods, dispatch] = React.useReducer(
-    methodsReducer,
-    INITIAL_STATE_METHODS
-  );
 
-  const cardTokenizeResponseReceived = (rest: TokenResult): void => {
+  const cardTokenizeResponseReceived = async (
+    rest: TokenResult
+  ): Promise<void> => {
     if (rest.errors || !props.createVerificationDetails) {
       props.cardTokenizeResponseReceived(rest);
       return;
     }
 
-    pay?.verifyBuyer(String(rest.token), props.createVerificationDetails());
+    const verifyBuyerResults = await payments?.verifyBuyer(
+      String(rest.token),
+      props.createVerificationDetails()
+    );
+
+    props.cardTokenizeResponseReceived(rest, verifyBuyerResults);
+    return;
   };
 
   // Fixes stale closure issue with using React Hooks & SqPaymentForm callback functions
@@ -92,78 +86,16 @@ const FormProvider: React.FC<ProviderProps> = ({ children, ...props }) => {
     cardTokenizeResponseReceived
   );
 
-  /**
-   * Helper function to update the state of the form and retreive the available methods
-   *
-   * @param methods The methods that you want to support
-   */
-  const methodsSupported = (
-    methods: MethodsSupported = props.methodsSupported || {
-      card: true,
-    }
-  ): void => {
-    let result = { ...INITIAL_STATE_METHODS };
-
-    for (const [method, enabled] of Object.entries(methods)) {
-      if (enabled) {
-        Object.keys(INITIAL_STATE_METHODS).map((initialMethod) => {
-          if (method === initialMethod) {
-            result = {
-              ...result,
-              [initialMethod]: 'ready',
-            };
-
-            return;
-          }
-        });
-      }
-    }
-
-    dispatch({
-      type: 'CHANGE_STATE',
-      // @ts-ignore
-      payload: result,
-    });
-  };
-
-  async function loadPayment(): Promise<void> {
-    methodsSupported();
-
-    await payments(applicationId, locationId).then((res) => {
-      if (res === null) {
-        throw new Error('Square Web Payments SDK failed to load');
-      }
-
-      setPay(res);
-
-      return res;
-    });
-  }
-
-  React.useEffect(() => {
-    if (applicationId && locationId) {
-      loadPayment();
-    }
-  }, [applicationId, locationId]);
-
-  if (!applicationId || !locationId) {
-    return <NoLocationIdOrAppId />;
-  }
-
-  if (!pay) {
-    return null;
-  }
+  if (!payments) return null;
 
   const context = {
-    ...methods,
-    formId: '',
-    payments: pay,
-    dispatchMethods: dispatch,
-    // @ts-ignore: Always true error
-    cardTokenizeResponseReceived: props.cardTokenizeResponseReceived
-      ? cardTokenizeResponseReceivedCallback
-      : null,
+    cardTokenizeResponseReceived:
+      // @ts-ignore: Always true error
+      props.cardTokenizeResponseReceived &&
+      cardTokenizeResponseReceivedCallback,
     createPaymentRequest,
+    formId: '',
+    payments,
   };
 
   return (