+++ src/index.ts
@@ -1,13 +1,1 @@
-import { SquarePaymentsForm } from './components/SquarePaymentsForm';
-
-export * from './components/AchPayButton';
-export * from './components/ApplePayButton';
-export * from './components/CreditCardInput';
-export * from './components/GiftCardInput';
-export * from './components/GooglePayButton';
-export * from './components/LoadingForm';
-export * from './components/NoLocationIdOrAppId';
-export * from './components/Separator';
-export * from './components/SquarePaymentsForm';
-
-export default SquarePaymentsForm;
+export * from './components';