+++ src/utils/isSSR.ts
@@ -1,0 +1,3 @@
+export const isSSR: boolean = !(
+  typeof window !== 'undefined' && window.document?.createElement
+);